Compact handshake records into less packets
Currently each record during the handshake process is sent as a
separate packet. The dTLS spec allows for combining records into a
signal packet so long as the packet fits within the MTU. Here we
implement this improvement, compacting records during the handshake
process into less packets. The number of packets sent for a complete
handshake before this change was 12. With this change the number is
reduced to 7. While improvements are likely hard to measure, this
should reduce tail handshake latency in lossy networks.